concatenated two strings

+++ 0x0B-malloc_free/2-str_concat.c
@@ -9,7 +9,7 @@
 
 char *str_concat(char *s1, char *s2)
 {
-	unsigned int lens1, lens2, i, j;
+	size_t lens1, lens2, i, j;
 	char *ptr;
 
 	if (s1 == NULL)
